fix: refactor models directory structure
