+++ src/models/food.py
@@ -1,7 +1,6 @@
 from datetime import datetime
-from typing import List, Optional
+from typing import TYPE_CHECKING, List, Optional
 
-from sqlalchemy import UniqueConstraint
 from sqlmodel import Field, Relationship, SQLModel
 
 from src.schemas.unit import FoodType, PrimaryUnit, UnitEnum
@@ -27,20 +26,6 @@ class FoodModel(SQLModel, table=True):
     product_db_id: Optional[int] = Field(index=True)
     created_at: datetime = Field(default_factory=datetime.utcnow, nullable=False)
 
+    if TYPE_CHECKING:
+        from .report import ReportModel
     reports: List["ReportModel"] = Relationship(back_populates="food")
-
-
-class ReportModel(SQLModel, table=True):
-    id: int = Field(primary_key=True, default=None, index=True)
-    food_id: int = Field(foreign_key="foodmodel.id")
-    user_id: int = Field(foreign_key="usermodel.id")
-    carbohydrate: Optional[float]
-    protein: Optional[float]
-    fat: Optional[float]
-    sugar: Optional[float]
-    energy: Optional[float]
-    reference: int = Field(default=0)
-    created_at: datetime = Field(default_factory=datetime.utcnow, nullable=False)
-
-    food: FoodModel = Relationship(back_populates="reports")
-    # __table_args__ = (UniqueConstraint("food_id", "user_id"),)
